upload_v2

+++ Server/src/main/java/Core.java
@@ -120,7 +120,7 @@ public class Core {
 
         public void checkCommand(ByteBuf command) throws Exception {
             System.out.println("r1 "+command.readableBytes());
-            //executorService.execute(()->{
+            executorService.execute(()->{
             while(command.isReadable()) {
                 try {
                     if (state == 0) {
@@ -130,6 +130,7 @@ public class Core {
                                 tmp = new byte[pathCDLen];
                                 command.readBytes(tmp);
                                 String pathChange = new String(tmp, "UTF-8");
+                                command.release();
                                 return;
                             case Patterns.UPLOADFILE:
                                 int pathDWFLen = command.readInt();
@@ -138,14 +139,11 @@ public class Core {
                                 fileName = new String(tmp, "UTF-8");
                                 if (Files.exists(Paths.get(fileName))) {
                                     System.out.println("File already exists");
-                                    while (command.isReadable()) {
-                                        command.readByte();
-                                    }
                                     //отправлять сообщение клиенту чтобы он не продолжал слать файл
                                 } else {
                                     fileSize = command.readLong();
-                                    countPack = fileSize / 256;
-                                    if (fileSize % 256 != 0) {
+                                    countPack = fileSize / Patterns.BYTESIZE;
+                                    if (fileSize % Patterns.BYTESIZE != 0) {
                                         countPack++;
                                     }
                                     System.out.println("count pack: " + countPack);
@@ -153,8 +151,9 @@ public class Core {
                                     System.out.println("filesize: " + fileSize);
                                     Files.createFile(Paths.get(fileName));
                                     readBytes = 0l;
+                                    state = 1;
                                 }
-                                state = 1;
+                                command.release();
                                 return;
                             case Patterns.DOWNLOADFILE:
                                 int pathUPFLen = command.readInt();
@@ -162,14 +161,20 @@ public class Core {
                                 command.readBytes(tmp);
                                 String pathDWF = new String(tmp, "UTF-8");
                                 //передача байтов самого файла
+                                command.release();
                                 return;
                             case Patterns.DELETEFILE:
                                 int pathDFLen = command.readInt();
                                 tmp = new byte[pathDFLen];
                                 command.readBytes(tmp);
                                 String pathDF = new String(tmp, "UTF-8");
-                                Files.delete(Paths.get(pathDF));
-                                System.out.println("File: " + pathDF + " deleted ");
+                                if(Files.notExists(Paths.get(pathDF))){
+                                    System.out.println("File: " + pathDF + " does not exist");
+                                } else {
+                                    Files.delete(Paths.get(pathDF));
+                                    System.out.println("File: " + pathDF + " deleted ");
+                                }
+                                command.release();
                                 return;
                             case Patterns.GETFILELIST:
                                 int pathFLLen = command.readInt();
@@ -181,17 +186,16 @@ public class Core {
                                 for (Object obj : dir2) {
                                     System.out.println(obj.toString().replace(root + "/", ""));
                                 }
+                                command.release();
                                 return;
                             default:
                                 System.out.println("Wrong command");
-                                while (command.isReadable()) {
-                                    command.readByte();
-                                }
+                                command.release();
                         }
                     } else if (state == 1) {
-                        System.out.println("HiState1");
-                        System.out.println("r1 "+command.readableBytes());
+                        System.out.println("input bytes "+command.readableBytes());
                         System.out.println(readBytes+"/"+fileSize);
+                        System.out.println(readBytes/(fileSize/100) + " %");
 //                        tmp = new byte[256];
 //                        while (readPack < countPack) {
 //                            if (readPack + 1 == countPack) {
@@ -216,19 +220,19 @@ public class Core {
                             }
                              command.readBytes(tmp);
                              readBytes += tmp.length;
-                            Files.write(Paths.get(fileName), tmp, StandardOpenOption.APPEND);
+                             Files.write(Paths.get(fileName), tmp, StandardOpenOption.APPEND);
                         }
                         if (readBytes.equals(fileSize)){
                             state = 0;
                             System.out.println("Upload file: " + fileName + " done");
                         }
+                        command.release();
                     }
-
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
             }
-            //});
+            });
         }
     }
 }
